+++ CHANGELOG.md
@@ -11,6 +11,24 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.0.0-rc.69 (2020-09-28)
+
+#### :boom: Breaking Change
+
+* Renamed `FieldGetter` -> `ValueGetter` `super/i-block/modules/field`
+
+#### :rocket: New Feature
+
+* Added `getter` for `set` and `delete` methods `super/i-block/modules/field`
+
+#### :bug: Bug Fix
+
+* Fixed a bug when a system field can't be watched after removal of a property `super/i-block/modules/field`
+
+#### :nail_care: Polish
+
+* Added more examples `super/i-block/modules/field`
+
 ## v3.0.0-rc.68 (2020-09-23)
 
 #### :boom: Breaking Change